load node/apps list before initial render
