+++ packages/playground/src/components/app-home/app-home/app-home.tsx
@@ -1,10 +1,7 @@
 import { Component, Element, Prop, State } from "@stencil/core";
 import { RouterHistory } from "@stencil/router";
 
-import AppRegistryTunnel, {
-  AppRegistryState
-} from "../../../data/app-registry";
-import PlaygroundAPIClient from "../../../data/playground-api-client";
+import AppRegistryTunnel from "../../../data/app-registry";
 import { AppDefinition } from "../../../types";
 
 @Component({
@@ -16,23 +13,16 @@ export class AppHome {
   @Element() private element: HTMLElement | undefined;
 
   @Prop() history: RouterHistory = {} as RouterHistory;
-
-  @State() apps: AppDefinition[] = [];
+  @Prop() apps: AppDefinition[] = [];
   @State() runningApps: AppDefinition[] = [];
-  @Prop() updateAppRegistry: (data: AppRegistryState) => void = () => {};
 
   appClickedHandler(e) {
     this.history.push(e.detail.dappContainerUrl, e.detail);
   }
 
   async componentWillLoad() {
-    this.apps = await PlaygroundAPIClient.getApps();
-
     // TODO: This is still mocked.
     this.runningApps = [{ ...this.apps[0], notifications: 11 }];
-
-    // Save to global state.
-    this.updateAppRegistry!({ apps: this.apps });
   }
 
   render() {
@@ -52,4 +42,4 @@ export class AppHome {
   }
 }
 
-AppRegistryTunnel.injectProps(AppHome, ["updateAppRegistry"]);
+AppRegistryTunnel.injectProps(AppHome, ["apps"]);
