+++ packages/playground/src/components/node-listener/node-listener.tsx
@@ -1,12 +1,11 @@
 declare var uuid: () => string;
 
-import { NetworkContext, Node } from "@counterfactual/types";
+import { Node } from "@counterfactual/types";
 import { Component, Element, Prop, State } from "@stencil/core";
 import { RouterHistory } from "@stencil/router";
 
 import AppRegistryTunnel from "../../data/app-registry";
 import CounterfactualNode from "../../data/counterfactual";
-import FirebaseDataProvider from "../../data/firebase";
 import { AppDefinition, WidgetDialogSettings } from "../../types";
 
 type NodeMessageHandlerCallback = (data: any) => void;
@@ -39,57 +38,6 @@ export class NodeListener {
   }
 
   async componentWillLoad() {
-    // TODO: This is a dummy firebase data provider.
-    // TODO: This configuration should come from the backend.
-    const serviceProvider = new FirebaseDataProvider({
-      apiKey: "AIzaSyA5fy_WIAw9mqm59mdN61CiaCSKg8yd4uw",
-      authDomain: "foobar-91a31.firebaseapp.com",
-      databaseURL: "https://foobar-91a31.firebaseio.com",
-      projectId: "foobar-91a31",
-      storageBucket: "foobar-91a31.appspot.com",
-      messagingSenderId: "432199632441"
-    });
-
-    const messagingService = serviceProvider.createMessagingService(
-      "messaging"
-    );
-    const storeService = {
-      async get(key: string): Promise<any> {
-        return JSON.parse(window.localStorage.getItem(key) as string);
-      },
-      async set(
-        pairs: {
-          key: string;
-          value: any;
-        }[]
-      ): Promise<boolean> {
-        pairs.forEach(({ key, value }) => {
-          window.localStorage.setItem(key, JSON.stringify(value) as string);
-        });
-        return true;
-      }
-    };
-
-    const addressZero = "0x0000000000000000000000000000000000000000";
-    const networkContext: NetworkContext = {
-      AppRegistry: addressZero,
-      ETHBalanceRefund: addressZero,
-      ETHBucket: addressZero,
-      MultiSend: addressZero,
-      NonceRegistry: addressZero,
-      StateChannelTransaction: addressZero,
-      ETHVirtualAppAgreement: addressZero
-    };
-
-    await CounterfactualNode.create({
-      messagingService,
-      storeService,
-      networkContext,
-      nodeConfig: {
-        STORE_KEY_PREFIX: "store"
-      }
-    });
-
     this.bindNodeEvents();
   }
